external api call and finding cleanest Region works, but only with an ugly workaround...
